refactor : Hand, Score state 패키지 안으로 이동
