+++ src/main/java/blackjack/domain/state/Blackjack.java
@@ -1,7 +1,5 @@
 package blackjack.domain.state;
 
-import blackjack.domain.participant.Hand;
-
 public class Blackjack extends Finished{
 
     public Blackjack(final Hand hand) {
