+++ src/main/java/blackjack/domain/state/Bust.java
@@ -1,7 +1,5 @@
 package blackjack.domain.state;
 
-import blackjack.domain.participant.Hand;
-
 public class Bust extends Finished {
 
     public Bust(final Hand hand) {
