+++ src/main/java/blackjack/domain/state/Hand.java
@@ -1,4 +1,4 @@
-package blackjack.domain.participant;
+package blackjack.domain.state;
 
 import blackjack.domain.carddeck.Card;
