+++ src/main/java/blackjack/domain/state/StateFactory.java
@@ -1,7 +1,6 @@
 package blackjack.domain.state;
 
 import blackjack.domain.carddeck.Card;
-import blackjack.domain.participant.Hand;
 
 import java.util.ArrayList;
 import java.util.List;
