+++ src/main/java/blackjack/domain/state/Stay.java
@@ -1,7 +1,5 @@
 package blackjack.domain.state;
 
-import blackjack.domain.participant.Hand;
-
 public class Stay extends Finished{
 
     public Stay(final Hand hand) {
