+++ src/test/java/blackjack/domain/state/HitTest.java
@@ -3,7 +3,6 @@ package blackjack.domain.state;
 import blackjack.domain.carddeck.Card;
 import blackjack.domain.carddeck.Number;
 import blackjack.domain.carddeck.Pattern;
-import blackjack.domain.participant.Hand;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
